Disable Google Maps
The GitHub repository visibility is now public in order to continue to
be able to build for free with Netlify. With that change, our Google
Maps API Key was exposed. I deleted that API Key from our Google account
so that it's no longer valid.

We're not using the Google Maps API anymore anyway so I also removed it
from the website config.